Bundle Updates
This is a bundle of updates that I have done some time ago and never pushed to the repos.

+ I have added a few new parts to the Discord Pack to add extra embed types to the message or embed message.
+ I have also added a request class Object that uses ObjectImp to build more Endpoints to interact with the same base system.
+ You also will find a new Example added, it uses the new Objects in the Discord Pack  such as a Discord Embed that has a Thumbnail, Footer, etc.

I need to do more testing on this commit before I decide this is the model that I would like to take for this kind of a request as it has many moving parts and many components that can be used.